py: move non-type-specific logic into Base, leaving subclasses as facades
Base now holds the AstRef and implements the ~28 operations that don't depend
on the concrete sort (op/args/variables/symbolic/concrete/annotations,
hash/repr/shallow_repr, canonicalize/identical/depth/is_leaf, simplify/replace,
and the annotation-management family). Bool/BV/FP/String inherit these and keep
only their typed operations (operators, conversions, concrete_value,
is_true/is_false, size, __reduce__). Each subclass retains its own inner field
for those methods and re-declares __hash__ (a class defining __eq__ without
__hash__ is unhashable in Python, even with an inherited one). Transform methods
return Bound<Base> via from_dynast, which yields the right subclass.

Verified against the Python test suite (maturin build + pytest): same results as
before the change.
